Tarea 07-Práctica en equipos/Ejercicio3/requirements.txt
numpy
Pillow
requests
threaded
